Salvando logado na sessão e campo usuario na tabela games
