+++ MyCollections/Areas/Identity/Pages/Account/Login.cshtml
@@ -31,9 +31,9 @@
                     <p>
                         <a asp-page="./ForgotPassword">Forgot your password?</a>
                     </p>
-                    @*<p>
+                    <p>
                         <a asp-page="./Register" asp-route-returnUrl="@Model.ReturnUrl">Register as a new user</a>
-                    </p>*@
+                    </p>
                 </div>
             </form>
         </section>
