+++ MyCollections/Views/Games/Index.cshtml
@@ -3,6 +3,9 @@
 <p class="margin_top">
     <a asp-action="Create">Novo jogo</a>
 </p>
+
+<button class="btn-default" id="btn_GetFromSteam" value="Atualizar informações do Steam"></button>
+
 <table class="table" id="tabela">
     <thead>
         <tr>
