+++ MyCollections/Views/Shared/_LoginPartial.cshtml
@@ -18,7 +18,7 @@
 else
 {
     <ul class="nav navbar-nav navbar-right">
-        @*<li><a asp-area="Identity" asp-page="/Account/Register">Register</a></li>*@
+        <li><a asp-area="Identity" asp-page="/Account/Register">Register</a></li>
         <li><a asp-area="Identity" asp-page="/Account/Login">Login</a></li>
     </ul>
 }
