+++ MyCollections/wwwroot/js/site.js
@@ -1,3 +1,10 @@
 ﻿$(document).ready(function () {
     $('#tabela').DataTable();
 });
+
+//function getFromSteam {
+//    $.ajax({
+//        method: "GET",
+//        url: "Games/GetFromSteam"
+//    })
+//}
